client-api/src/main/java/es/imaut/clientapi/controller/ClientController.java
package es.imaut.clientapi.controller;

import es.imaut.clientapi.dto.ClientDto;
import es.imaut.clientapi.service.ClientService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

import static org.springframework.http.ResponseEntity.ok;

@RestController
@RequestMapping("/clients")
@RequiredArgsConstructor
public class ClientController {
    private final ClientService service;

    @GetMapping
    public ResponseEntity<List<ClientDto>> findAll() {
        return ok(service.findAll());
    }
}
